ref: move codeowners endpoints into issues module
