+++ src/sentry/issues/endpoints/organization_group_search_view_details.py
@@ -12,7 +12,7 @@ from sentry.api.bases.organization import OrganizationEndpoint
 from sentry.api.serializers.base import serialize
 from sentry.api.serializers.models.groupsearchview import GroupSearchViewSerializer
 from sentry.api.serializers.rest_framework.groupsearchview import ViewValidator
-from sentry.issues.endpoints.bases import GroupSearchViewPermission
+from sentry.issues.endpoints.bases.group_search_view import GroupSearchViewPermission
 from sentry.issues.endpoints.organization_group_search_views import pick_default_project
 from sentry.models.groupsearchview import GroupSearchView
 from sentry.models.groupsearchviewlastvisited import GroupSearchViewLastVisited
