+++ src/sentry/issues/endpoints/project_codeowners_index.py
@@ -9,10 +9,11 @@ from sentry.analytics.events.codeowners_created import CodeOwnersCreated
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import region_silo_endpoint
-from sentry.api.bases.project import ProjectEndpoint
 from sentry.api.serializers import serialize
 from sentry.api.serializers.models import projectcodeowners as projectcodeowners_serializers
 from sentry.api.validators.project_codeowners import validate_codeowners_associations
+from sentry.issues.endpoints.bases.codeowners import ProjectCodeOwnersBase
+from sentry.issues.endpoints.serializers import ProjectCodeOwnerSerializer
 from sentry.issues.ownership.grammar import (
     convert_codeowners_syntax,
     create_schema_from_issue_owners,
@@ -20,11 +21,9 @@ from sentry.issues.ownership.grammar import (
 from sentry.models.project import Project
 from sentry.models.projectcodeowners import ProjectCodeOwners
 
-from . import ProjectCodeOwnerSerializer, ProjectCodeOwnersMixin
-
 
 @region_silo_endpoint
-class ProjectCodeOwnersEndpoint(ProjectEndpoint, ProjectCodeOwnersMixin):
+class ProjectCodeOwnersEndpoint(ProjectCodeOwnersBase):
     owner = ApiOwner.ISSUES
     publish_status = {
         "GET": ApiPublishStatus.PRIVATE,
